feat: add MPS support and ColSmolVLM support for local development
